requirements.txt
fastapi
uvicorn
langchain
langchain-openai
openai
PyGithub
python-dotenv
langfuse
requests